agent-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
redis
rq
